moved `get_avalon_database` and `set_io_database` to deprecated
